move OffsetMtr to its own package.
This allows us to use the same offsetMgr for both ingestion and
the clusterHandler
